tests/wpt/web-platform-tests/uievents/interface/click-event.htm

<!doctype html>
<html>
  <head>
    <title>Click event is a PointerEvent</title>
    <link rel="help" href="https://github.com/w3c/pointerevents/pull/317">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <div id='clicktarget'></div>
    <div id="log"></div>
    <script type="text/javascript">
    var clicktarget = document.querySelector("#clicktarget");
    var t = async_test("synthetic click event is a PointerEvent");
    clicktarget.addEventListener('click', t.step_func(function (e) {
      assert_equals(e.constructor, window.PointerEvent);
      assert_true(e instanceof window.PointerEvent);
      // Since this click is not generated by a pointing device, pointerId and
      // pointerType must have default values (0 and empty string)
      assert_equals(e.pointerId, 0);
      assert_equals(e.pointerType, "");
      t.done();
    }));
    document.querySelector('#clicktarget').click();
    </script>
  </body>
</html>
